--- a/test/TestUtilities/Test.Utility/SlowStream.cs
+++ b/test/TestUtilities/Test.Utility/SlowStream.cs
@@ -1,13 +1,11 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
 using System.IO;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 
-namespace NuGet.Protocol.Tests
+namespace Test.Utility
 {
     public class SlowStream : Stream
     {