I've created a project ``NuGet.Core.FuncTests`` which will run timeout stress tests. When profiling the current core unit tests I found that most of the time taken was from a few tests across the different projects. For netcore tests the runner currently doesn't do well with parallel assemblies so a few tests taking 10-30s meant pushed the total time to around an extra 90s.

Here is a rough estimation of the perf improvement:
Desktop unit tests 33s -> 16s
NETCore unit tests 150s -> 45s
Total saved: 122s

The new project takes 16s on desktop and 20s on core. Which gives 36s total. The reason this is so much faster is because the slow tests now run in parallel with each other in the same assembly and that works on the core xunit runner.

Since these tests cover timeouts/stress/downloads I think the func test section is a better place for them.

Any new tests like this should also go in this new general core func test project.